Добавил сохранение объектов энциклопедии, создал поле привилегий для пользователя

+++ greenserver/src/main/java/ru/alexey_ovcharov/webserver/logic/controllers/ImagesBean.java
@@ -12,7 +12,9 @@ import javax.persistence.PersistenceContext;
 import org.apache.commons.io.IOUtils;
 import ru.alexey_ovcharov.webserver.persist.Images;
 import ru.alexey_ovcharov.webserver.persist.ImagesForPlace;
+import ru.alexey_ovcharov.webserver.persist.ImagesForThing;
 import ru.alexey_ovcharov.webserver.persist.Places;
+import ru.alexey_ovcharov.webserver.persist.Things;
 
 /**
  *
@@ -23,7 +25,7 @@ import ru.alexey_ovcharov.webserver.persist.Places;
 public class ImagesBean implements Serializable {
 
     private static final String IMG_DIR = "C:/Storage/emerald-notes/greenserver/src/main/webapp/imgs";
-    
+
     @PersistenceContext(unitName = "ru.alexey_ovcharov_webserver_war_1.0PU")
     protected EntityManager em;
 
@@ -33,26 +35,52 @@ public class ImagesBean implements Serializable {
     public ImagesBean() {
     }
 
-    public String getImageForPlace(Places places) {
-        if (places != null) {
-            Places placesF = em.find(Places.class, places.getIdPlace());
-            Collection<ImagesForPlace> imagesForPlaceCollection = placesF.getImagesForPlaceCollection();
-            ImagesForPlace imagesForPlace = imagesForPlaceCollection.iterator().next();
-            Images idImage = imagesForPlace.getIdImage();
-            Images finded = em.find(Images.class, idImage.getIdImage());
-            try {
-                File file = new File(IMG_DIR + File.separator 
-                        + idImage.getGuid().toString() + ".png");
-                file.deleteOnExit();
-                try (FileOutputStream fileOutputStream = new FileOutputStream(file)) {
-                    IOUtils.write(finded.getImageData(), fileOutputStream);
+    public String getImageForThing(Things thing) {
+        if (thing != null) {
+            Things things = em.find(Things.class, thing.getIdThing());
+            Collection<ImagesForThing> imagesForPlaceCollection = things.getImagesForThingCollection();
+            if (imagesForPlaceCollection.size() > 0) {
+                ImagesForThing imagesForThing = imagesForPlaceCollection.iterator().next();
+                Images idImage = imagesForThing.getIdImage();
+                Images finded = em.find(Images.class, idImage.getIdImage());
+                try {
+                    File file = new File(IMG_DIR + File.separator
+                            + idImage.getGuid().toString() + ".png");
+                    file.deleteOnExit();
+                    try (FileOutputStream fileOutputStream = new FileOutputStream(file)) {
+                        IOUtils.write(finded.getImageData(), fileOutputStream);
+                    }
+                    return idImage.getGuid().toString() + ".png";
+                } catch (IOException ex) {
+                    ex.printStackTrace(System.err);
                 }
-                return idImage.getGuid().toString() + ".png";
-            } catch (IOException ex) {
-                ex.printStackTrace(System.err);
             }
         }
         return "";
     }
     
+    public String getImageForPlace(Places places) {
+        if (places != null) {
+            Places placesF = em.find(Places.class, places.getIdPlace());
+            Collection<ImagesForPlace> imagesForPlaceCollection = placesF.getImagesForPlaceCollection();
+            if (imagesForPlaceCollection.size() > 0) {
+                ImagesForPlace imagesForPlace = imagesForPlaceCollection.iterator().next();
+                Images idImage = imagesForPlace.getIdImage();
+                Images finded = em.find(Images.class, idImage.getIdImage());
+                try {
+                    File file = new File(IMG_DIR + File.separator
+                            + idImage.getGuid().toString() + ".png");
+                    file.deleteOnExit();
+                    try (FileOutputStream fileOutputStream = new FileOutputStream(file)) {
+                        IOUtils.write(finded.getImageData(), fileOutputStream);
+                    }
+                    return idImage.getGuid().toString() + ".png";
+                } catch (IOException ex) {
+                    ex.printStackTrace(System.err);
+                }
+            }
+        }
+        return "";
+    }
+
 }
